requirements.txt
numpy
scipy
matplotlib
pandas
tvb-library
environment.yml